ProtPlot

Documentation for ProtPlot.

ProtPlot.animate_molecule_dir

ProtPlot.animate_molecules

ProtPlot.animate_trajectory_dir

ProtPlot.ribbon_scene

ProtPlot.animate_molecule_dir — Method

animate_molecule_dir(export_path, input_dirs; kwargs...)

Reads all .xyz files in each input directory and animates them using animate_molecules.

ProtPlot.animate_molecules — Method

animate_molecules(export_path, molecules_cache; kwargs...)

Keyword arguments:

labels: A vector of labels for each molecule series.

frame_names: A vector of vectors of frame names for each molecule series.

rotation = 0.02: The rotation speed.

extra_seconds = 5: The number of seconds to show the extra frames after the animation.

end_rotation_speedup = 0.5: The end rotation speedup.

size = (1280, 720): The size of the figure.

framerate = 60: The framerate of the animation.

duration = 10.0: The duration of the animation.

theme = :black: The theme of the animation.

ProtPlot.animate_trajectory_dir — Method

animate_trajectory_dir(export_path, input_dirs; kwargs...)

Keyword arguments:

labels: A vector of labels for each input directory.

color_by: aa for amino acid color, chain for chain color, numbering for numbering color.

aa_colormap: A colormap for structures colored by amino acid.

chain_colormap: A colormap for structures colored by chain.

numbering_colormap: A colormap for structures colored by numbering.

rotation = 0.02: The rotation speed.

framerate = 24: The framerate of the animation.

end_rotation_speedup = 0.0: Additive end rotation speed.

end_ribbon_seconds = 3.0: The number of seconds to show the ribbon after the animation.

size: The size of the figure.

ProtPlot.ribbon_scene — Method

ribbon_scene(chain_backbones::AbstractVector{<:AbstractArray{T,3}}; backgroundcolor=:black, camcontrols=(;), kwargs...)

Render a protein as a ribbon diagram. The display will be automatically centered on the rendered ribbon, unless the user supplies camcontrols (see Makie's camera documentation for details).